# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MineOS Africa - Mining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all mine projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new mine project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a mine project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update a mine project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a mine project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project summary with aggregates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"
  /surveys:
    get:
      operationId: listSurveys
      tags: [surveys]
      summary: List all surveys
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of surveys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Survey"
    post:
      operationId: createSurvey
      tags: [surveys]
      summary: Create a new survey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSurveyBody"
      responses:
        "201":
          description: Created survey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
  /surveys/{id}:
    get:
      operationId: getSurvey
      tags: [surveys]
      summary: Get a survey by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Survey details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Survey"
        "404":
          description: Not found
    delete:
      operationId: deleteSurvey
      tags: [surveys]
      summary: Delete a survey
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List safety alerts
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, resolved, dismissed]
      responses:
        "200":
          description: List of alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create a new safety alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertBody"
      responses:
        "201":
          description: Created alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get an alert by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    put:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertBody"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
...
  /production:
    get:
      operationId: listProductionRecords
      tags: [production]
      summary: List production records
      parameters:
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of production records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionRecord"
    post:
      operationId: createProductionRecord
      tags: [production]
      summary: Create a production record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductionRecordBody"
      responses:
        "201":
          description: Created production record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRecord"
  /production/{id}:
    get:
      operationId: getProductionRecord
      tags: [production]
      summary: Get a production record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Production record details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionRecord"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        mineType:
          type: string
          enum: [open_pit, underground, quarry, alluvial]
        status:
          type: string
          enum: [active, inactive, exploration, closed]
        area:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - mineType
        - status
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        mineType:
          type: string
          enum: [open_pit, underground, quarry, alluvial]
        status:
          type: string
          enum: [active, inactive, exploration, closed]
        area:
          type: number
          nullable: true
      required:
        - name
        - mineType
        - status
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        totalSurveys:
          type: integer
        lastSurveyDate:
          type: string
          format: date-time
          nullable: true
        totalVolumeExcavated:
          type: number
        activeAlerts:
          type: integer
        totalProduction:
          type: number
        productionUnit:
          type: string
      required:
        - projectId
        - totalSurveys
        - totalVolumeExcavated
        - activeAlerts
        - totalProduction
        - productionUnit
    Survey:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        surveyDate:
          type: string
          format: date-time
        fileType:
          type: string
          enum: [dwg, dxf, shp, csv, orthomosaic, point_cloud, dem, geotiff]
        fileName:
          type: string
          nullable: true
        surveyMethod:
          type: string
          enum: [drone, gps, total_station, lidar, satellite]
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [processing, completed, error]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - surveyDate
        - fileType
        - surveyMethod
        - status
        - createdAt
    CreateSurveyBody:
      type: object
      properties:
        projectId:
          type: integer
        name:
          type: string
        surveyDate:
          type: string
          format: date-time
        fileType:
          type: string
          enum: [dwg, dxf, shp, csv, orthomosaic, point_cloud, dem, geotiff]
        fileName:
          type: string
          nullable: true
        surveyMethod:
          type: string
          enum: [drone, gps, total_station, lidar, satellite]
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [processing, completed, error]
      required:
        - projectId
        - name
        - surveyDate
        - fileType
        - surveyMethod
        - status
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
          enum: [slope_instability, deformation, seismic, flooding, equipment, other]
        status:
          type: string
          enum: [active, resolved, dismissed]
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - projectId
        - title
        - description
        - severity
        - category
        - status
        - createdAt
    CreateAlertBody:
      type: object
      properties:
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        category:
          type: string
          enum: [slope_instability, deformation, seismic, flooding, equipment, other]
        status:
          type: string
          enum: [active, resolved, dismissed]
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
      required:
        - projectId
        - title
        - description
        - severity
        - category
        - status
    UpdateAlertBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, resolved, dismissed]
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
...
    ProductionRecord:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        recordDate:
          type: string
          format: date-time
        material:
          type: string
        quantity:
          type: number
        unit:
          type: string
          enum: [tons, m3, kg]
        area:
          type: string
          nullable: true
        shift:
          type: string
          enum: [day, night, full]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - recordDate
        - material
        - quantity
        - unit
        - shift
        - createdAt
    CreateProductionRecordBody:
      type: object
      properties:
        projectId:
          type: integer
        recordDate:
          type: string
          format: date-time
        material:
          type: string
        quantity:
          type: number
        unit:
          type: string
          enum: [tons, m3, kg]
        area:
          type: string
          nullable: true
        shift:
          type: string
          enum: [day, night, full]
        notes:
          type: string
          nullable: true
      required:
        - projectId
        - recordDate
        - material
        - quantity
        - unit
        - shift
...